+++ CHANGELOG.md
@@ -21,6 +21,9 @@ with the `YY.MINOR.MICRO` scheme.
 - Default authentication method changed from `CHAP-SHA1` to `MD5` both for user creation and in connect CLI. This change affects new user creation and all system users (except the `pico_service` user), as a command-line interface of `picodata connect` and `picodata expel`. Also, default schema version at cluster boot is now `1`, not `0` as it was previously.
 Connection via `Pgproto` no longer requires additional manual step to change the authentication method. However if you use `iproto` the admin will have to manually change the authentication type.
 
+- New `picodata connect` and `picodata expel` argument `--timeout` for specifying
+the timeout for address resolving operation.
+
 - Support human numbers to configure memtx.memory.
   Supported suffixes: K, M, G, T, 1K = 1024
   (e.g picodata run --memtx-memory 10G)